+++ Firmware/shared/lib/utils.cpp
@@ -1,16 +1,17 @@
 #include "utils.hpp"
+#include <Arduino.h>
 
-Vector2D Vector2D::fromPolar(float angle, float length)
+Vector2D Vector2D::fromPolar(double angle, double length)
 {
     return Vector2D(cos(angle) * length, sin(angle) * length);
 };
 
-float Vector2D::length()
+double Vector2D::length()
 {
     return sqrt(x * x + y * y);
 };
 
-float Vector2D::angle()
+double Vector2D::angle()
 {
     return atan2(y, x);
 };
@@ -25,17 +26,17 @@ Vector2D Vector2D::operator-(const Vector2D &other)
     return Vector2D(x - other.x, y - other.y);
 };
 
-float Vector2D::operator*(const Vector2D &other)
+double Vector2D::operator*(const Vector2D &other)
 {
     return x * other.x + y * other.y;
 };
 
-Vector2D Vector2D::operator*(const float scale)
+Vector2D Vector2D::operator*(const double scale)
 {
     return Vector2D(x * scale, y * scale);
 };
 
-float determinant(Vector2D first, Vector2D second)
+double determinant(Vector2D first, Vector2D second)
 {
     return (first.x * second.y) - (first.y * second.x);
 };
